Add Quectel EC25 USB Serial Driver for Linux & Pi3
Add a recipie that patches the standard Yocto Linux Kernel and the
Raspberry Pi 3's kernel in a way that adds support for the Quectel EC25
LTE Module.  This allows the kernel to use the option kernel module to
drive the EC25, as it already does with the Quectel UC20.

This commit also updated the README.md file to document the
dependencies, usage and other details of this layer.