Add PPPd scripts
Add Point-to-Point Protocol Dialing scripts for use with the Point-to-
Point Protocol Daemon.

These scripts are based on the Quectel WCDMA & LTE Linux USB Driver User
Guide (V1.5) and are shared as part of this layer with their permission.

The specific scripts have been tailored for use in the UK's O2 network.
Comments in the quectel-chat-connect and quectel-ppp files indicated
where usernames, passwords and server details should be updated to match
other service providers.  In the future these scripts will be updated to
use enviroment variables for these times instead.